requirements.txt
# PDF Processing
PyPDF2>=3.0.0

# Microsoft Office Formats
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# HTML Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Markdown Processing
markdown>=3.4.0

# E-book Formats
ebooklib>=0.18
mobi>=0.3.3

# Progress Bar
tqdm>=4.65.0

# System Monitoring
psutil>=5.9.0

# Configuration
PyYAML>=6.0
